backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
cohere>=4.3.3
qdrant-client>=1.7.0
python-dotenv>=1.0.0
pydantic>=2.5.0
asyncio>=3.4.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
authlib>=1.3.0
requests>=2.31.0
cryptography>=41.0.7
itsdangerous>=2.1.2